Moved create_enemy functions to the separate enemies.py file fod code clearance
